Spells like listless power and dread touch should be taking down/blocking spells like augmentation and strenthen respectively.  The stacking code was being bypassed when a beneficial spell and a detrimental spell for the same effect were being compared.
With this code removed, the spells are compared and the stonger of the two
spells wins and replaces or blocks the other.

Without this change, for example, dread touch takes down strengthen on the
client, but the server still things strength is up, causing a mismatch.

With this change, client/server match.